feat: add middleware
