+++ backend/src/controllers/Pagination.py
@@ -1,7 +1,4 @@
-from functools import wraps
 from typing import List
-
-from flask import request
 
 
 class Pagination:
@@ -18,15 +15,3 @@ class Pagination:
             "total": self.total,
             "data": self.data,
         }
-
-
-def request_pagination(f):
-    @wraps(f)
-    def decorated(*args, **kwargs):
-        page = request.args.get("page", 1, type=int)
-        limit = request.args.get("limit", 10, type=int)
-        request.pagination = {"page": max(page, 1), "limit": max(limit, 10)}
-
-        return f(*args, **kwargs)
-
-    return decorated
